Components refactor and reorganize
