+++ frontend/src/modules/Nodes/components/RunningJob/RunningJob.tsx
@@ -1,137 +1,10 @@
-import React, { useState } from "react";
+import React from "react";
 import styles from "./RunningJob.module.scss";
-import { RunningNodeInfo, StateUpdate, StateUpdateObject, useNodesContext } from "../../context/NodesContext";
-import { CircularProgress } from "@mui/material";
-import { UpArrowIcon } from "../../../../ui-lib/Icons/UpArrowIcon";
+import { useNodesContext } from "../../context/NodesContext";
 import { SnapshotsApi } from "../../../Snapshots/api/SnapshotsApi";
-import { CheckMarkIcon } from "../../../../ui-lib/Icons/CheckMarkIcon";
-import { RightArrowIcon } from "../../../../ui-lib/Icons/RightArrowIcon";
 import BlueButton from "../../../../ui-lib/components/Button/BlueButton";
-import { EditIcon } from "../../../../ui-lib/Icons/EditIcon";
-import InputField from "../../../../DEPRECATED_components/common/Input/InputField";
 import { ErrorStatusWrapper } from "../../../common/Error/ErrorStatusWrapper";
-
-interface StateUpdateComponentProps {
-  key: string;
-  stateUpdateObject: StateUpdateObject;
-  runningNodeInfo?: RunningNodeInfo;
-}
-
-const StateUpdateComponent: React.FC<StateUpdateComponentProps> = (props) => {
-  const { key, stateUpdateObject, runningNodeInfo } = props;
-  const [runningUpdate, setRunningUpdate] = React.useState<boolean>(false);
-  const [parameterUpdated, setParameterUpdated] = useState<boolean>(false);
-  const [editMode, setEditMode] = useState<boolean>(false);
-  const [customValue, setCustomValue] = useState<unknown>(undefined);
-
-  return (
-    <div key={`${key}-wrapper`} className={styles.stateUpdateWrapper}>
-      <div>
-        {!runningUpdate && !parameterUpdated && (
-          <div
-            onClick={async () => {
-              if (
-                runningNodeInfo &&
-                runningNodeInfo.idx &&
-                stateUpdateObject &&
-                ("val" in stateUpdateObject || "new" in stateUpdateObject)
-              ) {
-                setRunningUpdate(true);
-                const stateUpdateValue = customValue ? customValue : stateUpdateObject.val ?? stateUpdateObject.new!;
-                const response = await SnapshotsApi.updateState(runningNodeInfo?.idx, key, stateUpdateValue);
-                setRunningUpdate(false);
-                setParameterUpdated(response.result!);
-                setRunningUpdate(false);
-              }
-            }}
-          >
-            <UpArrowIcon />
-          </div>
-        )}
-        <div className={styles.stateUpdateIconWrapper}>
-          {runningUpdate && !parameterUpdated && <CircularProgress size={32} />}
-          {!runningUpdate && parameterUpdated && <CheckMarkIcon />}
-        </div>
-      </div>
-      <div key={key} className={styles.stateUpdateComponentTextWrapper}>
-        <div className={styles.stateUpdateKeyText}>{stateUpdateObject?.key ? stateUpdateObject?.key.toString() : key.toString()}</div>
-        <div className={styles.stateUpdateValueText}>
-          {stateUpdateObject && (
-            <div className={styles.stateUpdateValueTextWrapper}>
-              {JSON.stringify(stateUpdateObject.old)}&nbsp;&nbsp;
-              <RightArrowIcon />
-              &nbsp;&nbsp;{JSON.stringify(stateUpdateObject.val ?? stateUpdateObject.new ?? "")}
-              <div
-                className={styles.editIconWrapper}
-                onClick={() => {
-                  setEditMode(true);
-                  setCustomValue(JSON.stringify(stateUpdateObject.val ?? stateUpdateObject.new));
-                }}
-              >
-                {!editMode && <EditIcon />}
-              </div>
-              {editMode && (
-                <InputField
-                  className={styles.newValueOfState}
-                  value={customValue as string | number | readonly string[] | undefined}
-                  onChange={(val) => {
-                    setCustomValue(val);
-                  }}
-                />
-              )}
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const GetStateUpdates: React.FC<{
-  runningNodeInfo: RunningNodeInfo | undefined;
-}> = (props) => {
-  const { runningNodeInfo } = props;
-
-  const handleClick = async (stateUpdates: StateUpdate) => {
-    const litOfUpdates = Object.entries(stateUpdates ?? {}).map(([key, stateUpdateObject]) => {
-      return {
-        data_path: key,
-        value: stateUpdateObject.val ?? stateUpdateObject.new!,
-      };
-    });
-    await SnapshotsApi.updateStates(runningNodeInfo?.idx ?? "", litOfUpdates);
-  };
-
-  return (
-    <>
-      {runningNodeInfo?.state_updates && Object.keys(runningNodeInfo?.state_updates).length > 0 && (
-        <div className={styles.stateTitle}>
-          State updates:
-          <div className={styles.updateAll}>
-            <BlueButton
-              className={styles.updateAllButton}
-              disabled={false}
-              onClick={() => handleClick(runningNodeInfo?.state_updates ?? {})}
-            >
-              Update all
-            </BlueButton>
-          </div>
-        </div>
-      )}
-      {runningNodeInfo?.state_updates && (
-        <div className={styles.stateUpdatesTopWrapper}>
-          {Object.entries(runningNodeInfo?.state_updates ?? {}).map(([key, stateUpdateObject]) =>
-            StateUpdateComponent({
-              key,
-              stateUpdateObject,
-              runningNodeInfo,
-            } as StateUpdateComponentProps)
-          )}
-        </div>
-      )}
-    </>
-  );
-};
+import { StateUpdates } from "../StateUpdates/StateUpdates";
 
 export const RunningJob: React.FC = () => {
   const { runningNode, runningNodeInfo, isNodeRunning, setIsNodeRunning } = useNodesContext();
@@ -218,7 +91,7 @@ export const RunningJob: React.FC = () => {
           {getRunningJobParameters()}
         </div>
       )}
-      <GetStateUpdates runningNodeInfo={runningNodeInfo} />
+      <StateUpdates runningNodeInfo={runningNodeInfo} />
       <ErrorStatusWrapper error={runningNodeInfo?.error} />
     </div>
   );
